Add particle-wall heat transfer parameters (#1538)
Description
This PR adds the wall parameters for particle-wall heat transfer, including the physical properties, the thermal boundary type (adiabatic or isothermal) and the temperature of the wall if it it isothermal. For now, only solid objects can have a temperature and will be used for particle-wall heat transfer.

Testing
Tests will be added in a future PR when particle-wall heat transfer is fully implemented.

Documentation
Documentation for lagrangian physical properties and solid objects was updated.