Add particle-wall heat transfer (#1534)
Description
This PR changes the particle_particle_heat_transfer files to calculate heat transfer for both particle-particle and particle-wall contacts with the same functions, as the calculation is very similar.

Testing
The test particle_wall_thermal_conductance is added to check the calculation of the thermal conductance for particle-wall contacts.