Add particle-wall heat transfer 3/3 (#1542)
Description
This PR finalizes the multiphysic DEM code by adding the calculation of particle-wall heat transfer and updating the temperature of solid objects when needed.
A parameter real_youngs_modulus is added, to be used for dem_mp when the Young's modulus is underestimated in the simulation.
A parameter disable_position_integration is also added to be able to freeze the position of particles. For example, this allows to have a higher time step for temperature variation in a previously loaded packed bed.

Testing
Multiphysic DEM test cases will be added in a future PR.

Documentation
Documentation for the new parameters is added.