Enable usage of the DOF Handler for the MatrixFree VANS solver (#1545)
Description
The Matrix Free VANS solver used to only be able to leverage functions directly into the assemblers. This has been fixed now and the actual interpolation of the void fraction is used within the assembly with a similar mechanism as to what is done for the buoyancy force.

Testing
All of the test results (except case 0) have changed (which was expected since the void fraction calculation has changed). The results have been updated accordingly.

Documentation
This is still highly experimental so this is not documented so far.
The next step will be the calculation of the particle-fluid forces within the matrix-free context.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>