Fix restart pvd checkpoint file in CFD_DEM simulations (#1552)
Description
When the Lagrangian post-processing tools were enabled, the simulation restart would try to read a wrongly named .pvd file for the lagrangian post-processing file. It did not have the right suffix.

Solution
Fixed the naming of the suffix.

Testing
Tested on one of the examples and it works well now.
---------

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>