Multiphysic DEM test cases (#1548)
Description
This example introduces the use of multiphysic DEM, with lined up particles that are heated by other particles (equilibrium.prm) or by a wall (wall-heating.prm). It verifies the models of particle-particle and particle-wall heat transfer with two test cases.
This PR also adds the dem-mp directory in documentation and examples.