Add piecewise polynomial transformation for the geometric interface reinitialization (#1546)
Description
The geometric redistanciation method had only a tanh transformation to go from the signed distance to the phase fraction. With it, it is difficult (technically imposible) to clamp the value of the phase fraction indicator to 0 and 1 away from the interface. To get close, the max redistanciation distance must be increased leading to higher computational cost.

Hence, a 4th degree piecewise polynomial transformation is added in this PR. It clamps the phase fraction to 0 or 1 at +/- the max redistanciation distance and it has a smooth change from 0 to 1.

The tanh and piecewise polynomial are implemented in the typical model class (such as the phase filter class).

A new parameter is added in the geometric interface reinitialization subsection: transformation type with 2 options (tanh and piecewise_polynomial).

This PR also improves the treatment of the level-set field and the iso-level in the VOF and SignedDistanceSolver implementation.

Testing
It has been tested with the rising bubble case and the applications tests were updated.

Documentation
The documentation will be added in my next PR along with the updated rising bubble example.