Add capability to calculate the void fraction from particles in the matrix-free VANS solver (#1556)
Description
The matrix-free VANS solver could only leverage a void fraction field derived from a function. This pull request adds the capability to use a void fraction derived from a particle field. The fluid flow is still not coupled with the particles (e.g. there is no drag force). The addition of the drag force will be a bit more complicated and will be the next step in this work.

Testing
A new test has been added that loads 10k particles inside of a cylinder and then simulates the flow through it. Again there is no drag, so the validation here is very approximate, but it will serve as a reference point.

Documentation
This is still a very experimental feature so it is not documented.


Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>