Add get_mapping and rotate_mapping functions (#1559)
Description
Reintroducing the mapping rotation function from #1536, this PR refactors the use of the mapping shared pointer by introducing a get_mapping() function. This way, the mapping_cache necessary for the rotor rotation can be updated and integrated with the code.

Testing
The rotate_mapping() function is already being tested by the test case introduced in #1536. Previous tests still pass after introducing the new get_mapping() function.

Documentation
The documentation for the rotor angular velocity, a parameter that is used for the mapping rotation, has been added in #1549.