Add surface area computation in VOF (#1562)
Description
The surface area computation of the interface was added for vof simulations when mass conservation is enable. For that, the NonMatching FEValues features of deal.II are used. Additionally, the geometric volume computation was changed to use the same features.

Testing
One unit test is added to check the surface and volume computations and their convergence rate in space. Both metrics converge to the second order.

Application tests outputting the mass conservation were updated because the additional surface area column was added to the outputted table.

Documentation
No documentation is required.