Update installation instructions for Rorqual (#1566)
Description
The installation instructions for the Digital Alliance clusters were a bit dated and convoluted, they also did not include Rorqual. This PR cleans up the documentation and also adds the flag recommandation for Rorqual. Furthermore, I have added a table to the main clusters we use with the link to their webpage to encourage people to read them.


Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>
Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>