Flag to delete old vtu files (#1569)
Description
This PR adds a feature that allows to delete old .vtu and .pvd files, when launching a simulation, before new ones are written.

Testing
The feature was tested with examples from all applications in which the flag -R is supported.

Documentation
The documentation has been added on a new section in Tools -> Command-Line Interface Options. Documentation regarding PR #1475 has also been added.