Unresolved CFD-DEM - Gas-solid Spouted Bed Example (#1564)
Description
This example is an updated version of the gas-solid spouted bed example in unresolved CFD-DEM, which simulates the spouting of particles in a rectangular bed. It is based on the simulation parameters from Geitani et al. The particle velocity magnitude results are compared to the experimental results from Yue et al.