Mortar: add application tests (#1575)
Description
This PR focus on a bug fix and two new application tests:
1- The center of rotation of the rotor was not being accounted for in the MortarManagerBase, which has been fixed.

2- Two new application tests for the mortar feature have been included, consisting of a MMS example. While one uses a dealii grid, the other uses a gmsh file.

Solution
The center_of_rotation is included in the point_to_angle computation, as well as in the MortarManagerBase. It is being read in the Mortar section of the prm file.

Testing
The application tests shows expected convergence for a Q2-Q2 example.