Matrix free default parameters (#1572)
Description
The matrix-free solver default parameters does not use the eigenvalue estimation for the relaxation coefficients. I don't think this is a good practice because the eigen value estimation for the relaxation coefficient has been shown to work very well. This PR changes the default value of that parameter.

Testing
All the test pass, but i changed 2 of the tests to not use the eigenvalue. These are 3D MMS test with Q1 which are not very good with inverse diagonal so I am not surprised.

Documentation
The documentation has been changed to reflect the new parameters.