Example 3d cylinder (#1570)
Description
This PR introduces a new example which is the 3D turbulent flow around a cylinder at a Reynolds of 3900. This PR makes an initial introduction of the example and focuses on the validating the C_D, St and the pressure distribution around the cylinder. I plan to make another PR in the future to add additional comparisons and validation as I gather more results. However, I feel like the PR already contains a lot of validation for the example to stand-off on its own.


Signed-off-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>