Deprecate gas-solid-spouted-bed example (#1578)
Description
Since the introduction of the gas-solid-spouted-rectangular-bed the old gas-solid-spouted-bed example is not relevant anymore. This PR deprecates this old example and removes it from the documentation in the spirit of continuous improvement of the examples we have.

Documentation
All part of the documentation related to that example have been removed.


Co-authored-by: flo-op <florence.caron02@gmail.com>