Mortar: add application test with rotated mesh (#1579)
Description
The rotor mesh rotation was not in the right place in the code, so that at the start of the simulation the accessed mapping information was still incorrect. This PR fixes this, and adds an application test for a rotor rotated mesh.

Solution
The rotate_mapping() is now done at the beginning of the simulation, as well as when the mesh is refined (called inside setup_dofs()).

Testing
A new application test of MMS with mortar and a rotated rotor mesh has been added.