Dynamic particle insertion in CFD-DEM (#1576)
Description
This PR adds the possibility to insert particles in CFD-DEM simulations. It uses the different insertion methods from DEM. The functions set_insertion_type() and insert_particles() are replicated from dem.cc to the CFD-DEM coupling file and adapted to the existing code for the CFD-DEM coupling. The insert_particles() function is called after the last DEM iteration in the solver to avoid solving for particles that do not yet feel the effect of the fluid.

Testing
The new test dynamic_particle_insertion tests the insertion of two particles in a CFD-DEM simulation, as well as testing the flow of the particles and the fluid. There are no particles in the beginning since no DEM files are read prior. The parameter void fraction time derivative is set to false in the cfd-dem subsection to avoid high derivatives from the insertion.

Documentation
There are no new simulation parameters. You simply need to add the DEM insertion info to the CFD-DEM parameters file.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>