Cfd dem void fraction linear solver (#1580)
Description
The calculation of the void fraction within the VANS and the CFD-DEM solvers can require the solution of a linear system of equation if the void fraction is calculated from particles.
The tolerance for the calculation of this void fraction was hard coded to 1e-15 and the parameters for the linear solver could not be controlled independently from those of the navier-stokes equations. This PR fixes that

Testing
Since the default value now does not match the previous one, all of the examples and tests were modified so that the tolerance of 1e-15 appears explicitely instead of implicitely.

Co-authored-by: mivaia <59896845+mivaia@users.noreply.github.com>
Co-authored-by: flo-op <florence.caron02@gmail.com>