Fix warnings introduced by functions now marked for deprecation in deal.II 9.8-pre (#1582)
Description
The introduction of the deal.II-9.8-pre version has marked some functions for being deprecated in the future.
Furthermore, I have found a small bug with related to error estimation with multiple variables. When establishing the coarsen and the refine flag with multiple variables, a loop was made over the refine vector, but the index of this loop was made to calculate both the coarsen and the refine vector. This inherently assumes that the refine and the coarsen vector have the same length, which is not a given. I have thus used two loops instead.
Finally, the matrix_free_non_linear prototype was not running in parallel anymore since the solution within the multigrid was not adequately ghosted.

Solution
I have fixed all of the warnings related to 9.8-pre
I have changed the looping logic to use two loops with the correct size instead of assuming both vectors had the same size
I have updated the ghost variables for the mg_solution vectors at every level.
Testing
No tests should be changed by this procedure. However, this might "destroy" the deal.II 9.6 CI. Since deal.II 9.7 is very close to being released, if this is an issue I will just disable the 9.6 warning CI and prepare to introduce the 9.7 warning CI with the release of 9.7.

Documentation
Nothing to document.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>