Allow compatibility with 64bit indices (#1585)
Description
The current version of Lethe does not compile with 64 bit indices because of a few errors where dof indices were assumed to be unsigned int.

Solution
Fix this by using the appropriate type where needed.

Testing
All test also pass in the 64bit indices version, so this is cool

Documentation
If 64 bit indices work well on the clusters for very large stuff, I will make another PR with the documentation on how to enable them.