Remove checks to deal.II versions which are not supported anymore (#1588)
Description
In the past, we had to introduce numerous checks for deal.II versions, notably when we introduced the parallel particle architectures. These tests check if the deal.II version if superior to version 9.3 or 9.4, which are versions that we do not even support anymore. Consequently, this PR removes these checks and only keeps the one which are related to deal.II 9.6 and 9.7. Once the official deal.II 9.7 release is finished, I will also remove support to deal.II 9.6 and move support to deal.II 9.7 as a minimum

Testing
This refactoring does not include new tests. Old tests should still pass as if nothing has changed.

Documentation
This does not require any modifications to the documentation.


Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>