Improve application documentation (#1587)
Description
The documentation of the different Lethe applications was nice, but i feel it does not adequately explain which solver can be used for what. For example, the lethe-fluid-matrix-free solver should be used for high-order simulations or pretty much every single-phase simulations, but this is not explained anywhere and may give a wrong impression to first time users.

I have tried to come up with an improved version of the table which explains what can be used where to give a better explanation of the different solvers in the hope that this gives a clear first impression.

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Justin Lamouche <135653039+Julamipolymtl@users.noreply.github.com>