Added the same check as the CFD-DEM (#1593)
Description
In #1576 , the default insertion frequency was changed to 0. When not defined in a prm file (like for the rotation drum example, which uses two prm files with a restart), a division by zero would occur with the modulo.

This wasn't a problem in the CFD-DEM, since a "if" statement verifying if the insertion frequency was equal to zero was added.

Solution
Added the same "if" statement in the DEM.

Testing
The rotation drum example works now.

Documentation
N/A