Particle-wall collision statistics logging in DEM (#1586)
Description
This PR adds a feature which allows to log particle-wall collisions statistics. It is integrated in the current calculate_particle_wall_contact() function. The log is created at the end of the simulation and contains, for each collision event, the id of the particle and the id of the boundary it collided with, the simulation time at the start and the end of the collision as well as the velocity and the angular velocity at the start and at the end of the collision.

Testing
A new application test was created where the start and end of each collision is printed in the output file.

Documentation
The following parameters are introduced under a subsection called Particle-Wall Contact Statistics Logging in Model Parameters.

enable particle wall collision statistics: Enables the feature. Default is false.
collision statistics file name: Name of the file of the log.
verbosity: Allows to print the start and the end of collisions in the terminal.
log collisions with all walls: Allows to log the particle-wall collisions with all walls. Default is true. To log the collisions with specific walls/boundaries, set to false.
wall boundary ids: Boundary ids of the walls if log collisions with all walls is set to false.

Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>