Dg shock capture (#1595)
Description
The DG implementation of the tracer was not very adequate for pure advection if the DG order was larger than 0 (so DG Q1 and above). This is because we did not have any sort of limiter in place. This PR adds a Moe limiter (see documentation for reference) which is an a posteriori limiter that works quite well for scalar equations. It is not very invasive and not very expensive and I have managed to get very adequate results with it.

Testing
The DG advection test has been changed to use the limiter and I have added a new test with BDF2 which really requires the limiter to get adequate results.

Documentation
Since the limiter is generic for all scalar equations, I have documented it in the stabilization section where I believe it adequately belong.

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>