Clarify documentation for deal.II 9.6 - Add two warnings for apt installs (#1600)
Description
The current master version of Lethe does not support deal.II 9.6 anymore, but we do not discuss this in the documentation anywhere. This means that our current installation instruction that use the "apt" install on both Linux and Windows just do not work if people use them. This PR adds a warning that specifies that the deal.II 9.6 version should only be used with the latest stable release of Lethe ( v1.0.1 which is our latest release)