Update first simulation documentation (#1603)
Description
I did not like the first simulation part of the documentation. It teaches user to work in a very weird fashion in which they do not install the executables but either copy the executables or run them directly from the build folder. This is not a typical linux fashion. I have streamlined the documentation to focus on people installing, but I still note the other alternatives.