Sort particles after insertion in CFD-DEM (#1604)
Description
This PR fixes a bug that occurred when inserting particles in CFD-DEM. A segmentation fault would occur when using the QCM void fraction method for a simulation run on multiple cores.

Solution
A call to the sort_particles_in_subdomains_and_cells is added to the insert_particles function to ensure the particles were properly sorted directly after their insertion.

Testing
The problematic simulation case was run to ensure the solution fixes the issue. Since the position of insert_particles in the simulation loop was changed, the output files for the collision_log application test and the dynamic_particle_insertion application test were changed. The particles insertion now occurs before the calculation of the void_fraction.