Initialize VOF using geometric redistanciation (#1602)
Description
The VOF model initial condition had a projection step capability to generate a smooth initial condition. However, it is also interestsing to be able to define an initial condition using the geometric redistanciation. This PR adds this capability. I also took the opportunity of moving the implementation of the initial conditions class to the .cc

Testing
All tests results remain unchanged. I am not sure a new test is required, but if you think it's necessary I can add one.

Documentation
The new parameter is discussed in the documentation.

Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>