Make Lethe work adequately without metis and fix some issues with null pointers (#1610)
Description
Although Lethe does not require METIS, the solver will hang if a simulation is ran in parallel with a simplex triangulation. This PR addresses this issue by adding correct assertion.
Furthermore, Lethe struggles to compile on GCC 15.+, due to some nullptr that are not necessary.

Solution
Check if deal.II is actually compiled with METIS, if not, check if Zoltan is available, if not, throw with an adequate error message
Replace the nullptr with the appropriate mpi_communicator for the distributed triangulations used in the box refinement.
Testing
All tests should pass as usual.

Documentation
Does not modify anything in the documentation