Mortar: fix update function (#1614)
Description
The mortar update was not accounting for the updated constraints when rotating the mesh, and the initial condition was being erased when updating the mortar information.

This PR also changes the way of reading msh files, so that two separate rotor/stator meshes are taken as input (the same way it was being done for dealii grids).

Solution
Instead of calling setup_dofs() inside the update_mortar_configuration(), the update is done only in what is necessary: rotating the rotor mapping, defining constraints, reinit mortar operators. and update the sparsity pattern.

Testing
No new tests have been added, but previous ones have been updated after the bug fix.