Output slice prototype (#1612)
Description
This prototype tests the DataOutResample class of deal.II, which can be used to output slices of your dataset instead of its entirety. The steps include generating a new triangulation and resampling the results in the intersection points between the new and the original ones.

Testing
The solution seems to work in the canonical case I have tested. In the future, when the final version will be implemented, the prototype will become a test.

Documentation
No need for the time being.


Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>