Collision stats logging with multiple MPI processes (#1613)
Description
This PR modifies the write_collision_stats function to allow the logging of particle-wall collisions when running simulations using multiple MPI processes. Each rank has its own CompletedCollisionLog with the particle-wall collisions that occured, and they must write to the same file one at a time to avoid corrupting the data.

It also adds two small modifications. A dem_current_time is defined to log the collision start and end time in CFD-DEM simulations. This allows to have a start time different from the end time in case a collision starts and ends in a single CFD iteration. The particle diameter and mass are added to the log.

Testing
No new test.

The solution was tested by running a simulation on 12 MPI ranks. Each rank would first add a line that announced it would start writing the data to the file, and then add a line that announced it was done writing data. All 12 ranks wrote data in order, one after the other, with no overlap.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>