Clean assertions (#1620)
Description
In many locations in the core library we were using AssertThrow where a simple Assert could have been used. The Assert macro is actually removed when compiling in Release, which is not the case with AssertThrow. The AssertThrow macro should only be used to assert parameters for example.

Testing
All tests should remain the same.

Documentation
Nothing to do here.