SDIRK scheme implementation (#1598)
Description
This pull request implements the SDIRK (Singly Diagonally Implicit Runge-Kutta) time integration method for the unsteady Navier–Stokes equations.

The current implementation is quite brute-force and would surely benefit from a proper refactoring to improve efficiency and readability. The idea is to also use the SDIRK scheme for other applications beyond just lethe-fluid.

The code is structured to match as closely as possible the existing BDF structure. Here's how it works:

At each time iteration, we loop over the stages. For "traditional" schemes (mainly BDF), the loop is straightforward: there is only one stage.
At each stage, for the SDIRK scheme, we compute the vectors kᵢ.
However, the structure of Lethe is designed to solve the nonlinear system on the velocity. So instead of directly solving kᵢ = F(tᵢ, uᵢ), we solve the equivalent system:
(uᵢ* - uₙ) / (h aᵢᵢ) - ∑_{j=0}^{i-1} (aᵢⱼ / aᵢᵢ) kⱼ = F(tᵢ, uᵢ)

Once we have computed the 
u
i
∗
 vector, we store the associated 
k
i
 vector in order to compute the next stages and the final sum.
At the end of the loop, the solution is computed as:
uₙ₊₁ = uₙ + h ∑ bᵢ kᵢ

The main changes are implemented in the solvers/navier_stokes_base.cc file, within the iterate() function.
The Butcher table is stored in scratch-data because it's required during matrix assembly.

Testing
The code has been tested on the lethe-fluid/application-test Taylor–Green vortex (GLS formulation) and yields results close to the BDF2 scheme.
The test outputs, at each time step, the enstrophy, the kinetic energy, and the L2 error.

The goal is now to test this new scheme on multiple examples and with different Butcher tables.

Documentation
The Butcher table is set in the scratch-data.
The variable sum_over_previous_stages is now required for matrix assembly.

Co-authored-by: etiennemarin <etienne.marin@ensta.fr>