Enhance geometric redistanciation performances (#1621)
Description
The geometric reinitialization method was really expensive in terms of computational time. This PR improves the performances by reducing the number of cells in the layer's loop (only the cells in a corona around the interface are visited) and the number of calls to FEPointEvaluation.reinit().

Testing
All the tests are still successful.

Documentation
No need for new documentation

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>