Refactor output results into struct (#1624)
Description
The goal of this refactor was to separate the write_output_results function into two: one that gathers the results in a single container (vector of structs) and another that takes the struct data and writes the vtu outputs. This is a step before adding a new feature that will allow results to be output in a mesh different than the one of the simulation. This can be used, for example, to output only a slice of the domain instead of the entirety of the results. With the gathering of the results and the instantiation of the feeding of the data_out object being done in two different functions, other instances of DataOut can be generated.

This PR also deprecates the smoothers of post-processing fields.

Testing
Test cases included the example case flow around cylinder, a simple case for the tracer in a micromixer, and the validation cases launched on Lucille.

Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>