Fix bugs with constraints when using local smoothing (#1626)
Description
The Taylor-Couette transient case was not converging when using local smoothing.

Solution
Several bugs were identified and fixed:

Pressure constraint was not correctly set for local smoothing.
A bug in the periodic boundary conditions was introduced since an additional boundary type periodic_neighbor was introduced in Use map to store boundary conditions instead of vectors #1330.
A timer that was throwing an error in debug mode.
Testing
A minimal Taylor-Couette test with periodic boundary conditions was added to ensure that this does not happen again when using lsmg and gcmg as preconditioners for the matrix-free application.

Documentation
It does not change anything regarding the documentation.



Co-authored-by: Peter Munch <peterrmuench@gmail.com>