Refactor VoidFractionManager to become ParticleProjector and enable projection of other fields (#1618)
Description
In the current work towards matrix-free CFD-DEM, it becomes relevant to be able to project particle-laden information onto the CFD mesh. This PR enables this features by creating a new mechanism that can project an arbitrary particle-laden field onto the CFD mesh.
Since the void fraction manager is not just for void fraction anymore, this class has been renamed to become the particle projector, which essentially projects particle information onto the CFD mesh.

Testing
This is a bit difficult to test with a meaningful test. I will work in future PR to add unit tests for theses functionnalities instead of relying on simulation. This feature is not enabled by default and I don't plan on exposing it to the users for a while, so I'm not overly worried about this.

Documentation
The particle velocity projection parameter is documented.

Co-authored-by: OresteMarquis <83602443+OresteMarquis@users.noreply.github.com>