Cfd dem vof - variation of fluid properties with vof (#1560)
Description
This PR adds the capability of taking into account the variations of the density and viscosity of the fluid with the volume phase fraction computed with the VOF scheme. These properties are then interpolated at the particles locations when calculating the forces involved in the VANS equations.

Testing
This feature has been tested on the example of the single particle sedimentation with buoyancy and drag forces included. A prm file is used where the VOF phase fraction is initialized. A high density and a low density fluid are also used. The VOF phase fraction is set to 0 in one fluid, and to 1 in the other. The results obtained with and without VOF are identical for the same fluid.