Fix and streamline the particle-fluid assemblers (#1633)
Description
Some of the particle-fluid force constructor were taking into argument parameters that they did not require such as the full lagrangian particle properties objectif. This makes it really invasive for the matrix-free stuff I am trying to achieve. This PR cleans this up.

Testing
All the tests should be non-affected by this.

Documentation
Nothing to document here. I took the opportunity to improve the doxygen of this part of the code.


Co-authored-by: Justin Lamouche <135653039+Julamipolymtl@users.noreply.github.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>