Cleanup scratch data for particle-fluid interaction and add defensive assertions (#1634)
Description
There were some issues with the scratch data for particle-fluid interaction:

Many functions were in the .h when they could have easily been in the .cc -> This has been fixed by moving them to the .cc
There are many locations were assumptions that a reinit is only called if the physics has been enabled were never asserted for. This is not a good defensive programming practice -> I have added a bunch of assertions whenever an enable is required or assumed
For a weird reason, gather the fluid information at the particle location required that the fluid information would have already been reinit and gathered. This meant that we could not gather the particle information without gathering the fluid information first, which is a fallacy since both are independent (they use different quadrature etc.) -> This has been fixed by making sure that the velocity_cell is used instead of the fe_values.get_cell() function.
The fluid properties array at the particle location was being resized at every iteration. This is not required for this array. The only reason why the other array are being resized is because of a weird quirk in fe_values that requires that the array used for the calculation be of the same size as the number of point in the quadrature.
Testing
No tests should be changed by this.

Documentation
No documentation change