Create random number container moved to utilities (#1639)
Description
The create_random_number_container function, which is currently being used when using the volume insertion in DEM, is moved to the utilities.h file.

Is changed is required for the particle ray-tracing PR, which also needs to create random number when inserting photons. This way, no code is duplicated.

Testing
Since this PR only changes the location of a function, no new test is required IMO. Otherwise, every old tests are passing.

Documentation
N/A