Particle force projection qcm (#1642)
Description
This PR introduces the capability to project the particle-fluid forces onto the CFD mesh using the ParticleProjector functionalities. This is an experimental feature that is not usable from the outside. It not tested, but I just want to break the development of this feature into chunks so I can start writing tests on one side while I also continue working on it as part of the matrix-free framework.

Testing
There are no test for this feature at the moment. I will write unit tests once I have realized if this feature is sufficient (or not) for the matrix-free solver.

Documentation
Not a usable feature, so not documented