Dem rolling resistance naming convention (#1640)
Description
The naming convention related to rolling resistance models is redundant. When setting the type of model :

"set rolling resistance torque model = constant_resistance "

The "_resistance" is not necessary. This PR is resolve #1420 .

Testing
Every test are passing and I tested that the warning message informing the user that he is using a soon to be deprecated parameter naming convention appears when using the old versions.

Documentation
The documentation was updated using the new convention.