Add simulations relaunch script for cluster usage (#1636)
Description
When launching multiple simulations on the cluster it is easy to loose track of which simulations are COMPLETED successfully and which ones have rather TIMEOUT or FAILED. This PR introduces a new script to the utilities of Lethe that allow to check the status of a list of simulations and relaunch them if desired.

The script goes through the different simulation folders and regroups the TIMEOUT and FAILED jobs in output files (timeout_simulations.txt and failed_simulations.txt) for the user to be able to investigate the state of the simulation at the end of the jobs before necessarily relaunching them. After investigating if there is an issue or not, the user can call again the script with appropriate flags to be able to relaunch the simulations that are not COMPLETED.
The script can relaunch the simulation from a checkpoint without having to manually modify the prm files of requested simulations.

A detailed description on how the script works and the arguments can be found in the script itself.

Testing
The script was tested numerous times on Rorqual when relaunching cases for a 3D rising bubble study.