Fix checkpointing (#1652)
escription of the Problem
The checkpointing with TableHandler tables had some issues. When checkpointing and restarting, the tables were serialized and deserialized on all ranks. Since the tables were only owned on rank 0, this resulted in an overwriting of data in the tables, and data history was lost at restart.

Solution
The bug was fixed by adding two arguments to the serializing and deserializing functions:

The MPI communicator
A boolean indicating if the serialization/deserialization must be done on 1 rank (rank 0); this is set to true by default.
If wished, in the future, with the current changes, it is also possible to serialize/deserialize on multiple cores, by setting the default parameter serialize_on_rank_zero/deserialize_on_rank_zero to false. Some slight modifications will be need to the functions to ensure correspondence between checkpoint files and processors.

Testing
Tested with a 3D rising bubble case on Rorqual.
Waiting for CI