Switch the website to a new branch
The website was currently hosted on a branch called gh-pages-v2.
In the past, we have been uploading extensive artifacts to that branch
from the feature branches which have in turn increased the size of the
repository. A previous PR (#1654) has fixed this by only pushing the
artifact on the master branch. Now to decrease the size of the repo we
refresh the webpage to use a new branch in order to remove the previous
branch which has grown too big to clone.