Improve rayleigh taylor example (#1637)
Description
The Rayleigh Taylor example was a bit outdated. It still only used the projection algorithm for redistanciation.
This PR improves the example by using the geometric method. I also improved the visuals, the movie and etc. according to our new standards. I really like the example now.

Testing
Not much it's just an example

Documentation
It's really a documentation PR :)


Co-authored-by: Bruno Blais <bruno@ivtm210.ads.uni-magdeburg.de>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>