Add missing space (#1656)
Description
A small bug was introduce at the last commit of PR #1636. A missing space in the script did not allow the user to change the restart to true for failed simulations.

Solution
The missing space has been added.

Testing
Tested with a failed 3D rising bubble simulation on Rorqual.