Hotfix sharp output (#1658)
Description
The sharp-edge solver used in resolved CFD-DEM would crash at the first VTU output produced. This is because no dof handler was being attached to the output hook.

Solution
This PR temporarily fixes that by attaching the dof handler of the main problem. A better solution will be implemented by leveraging the full mechanisms that @voferreira implemented with the struct. @voferreira I'll let you do that in a follow-up PR.

Testing
Tested on any of the examples and now they run.